+++ packages/core/crdt_socket_sync/lib/src/relay/server/relay_client_session.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 
 import 'package:crdt_lf/crdt_lf.dart';
 import 'package:crdt_socket_sync/src/common/common/common.dart';
+import 'package:crdt_socket_sync/src/common/common/utils.dart';
 import 'package:crdt_socket_sync/src/common/server/client_session.dart';
 import 'package:crdt_socket_sync/src/common/server/client_session_event.dart';
 import 'package:crdt_socket_sync/src/relay/common/common.dart';